Enabling alternative controller rest API classes in preprocess
